move k8s client under utils
