+++ app/utils/k8s/services_test.go
@@ -13,7 +13,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/fake"
 
-	"github.com/cloudzero/cloudzero-agent/app/k8s"
+	"github.com/cloudzero/cloudzero-agent/app/utils/k8s"
 )
 
 func TestGetKubeStateMetricsURLByName(t *testing.T) {
